VDI.export_changed_blocks: only require SR of vdi_to
If the VDIs are from different SRs, then the changed blocks are all the
blocks of vdi_to, so the SR of vdi_from doesn't have to be connected.
Using SR.forward_multiple_op caused a bug where the
export_changed_blocks call was forwarded to a slave, even though this
call has to be run from the SR master, which is the pool master in case
of a shared SR. The forward_vdi_op function correctly forwards the call
to the SR master.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>