CA-260352: Fix upgrade from 1.14-ppx to 1.30
To avoid problems when upgrading to the already-released 1.30 version of xapi,
we need to ensure the database doesn't contain any new fields. Therefore this
commit removes the new field 'enforce_homogeneity' from the database. However,
we do need the functionality the field provided, and hence we're moving it to
other_config. Unfortunately the object in question, 'Pool_update', doesn't have an
other_config field, hence we're storing it on the associated Pool_patch
record, which does have the field.

The other_config key will be:

    "enforce_homogeneity"

which will be set to "true" or "false".

In the hopefully vanishingly small chance that we can't find the field (ie,
the pool_patch object has been removed or the other_config map has been
blanked), we will assume the value would have been 'true', unless a pool-wide
override has been set.

In future, this functionality will be removed in favour of a the first-class
field on the pool_update object.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>