Pool_patch: remove enforce_homogeneity from other_config when syncing update
In resync_host we read updates from files and update the xapi database.
An update might already be in the database. Previously an update's
enforce_homogeneity flag was stored in its corresponding patch object
but we now store it in a field of an update directly. To avoid
confusion, we delete the entry in the pool patch object.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>