Implement VDI.get_nbd_info XenAPI call
Forward the call to master, because this call only does database
accesses, but it does many of them, so it's more efficient to do this on
the master.

I have skipped empty IPs, because the IPs can be temporarily empty if
they are configured via DHCP. And PIFs may not even have an IP, in which
case it is the empty string again.

In order to correctly format the IPv6 addresses in the output URIs and
conform to the RFC3986 specification of URI syntax, which says that IPv6
addresses should be enclosed in square brackets, I had to construct the
URIs by hand. I could not use the Uri library, because it would
incorrectly escape the colon in the IPv6 addresses using a URL escape
code.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>