CP-23497: Make VDA API calls default to `VM Operator` role
The guidelines for choosing the RBAC role for new API calls
state that `VM Admin` is required if the call is "related to
creating, destroying, cloning, importing/exporting VMs"; since
VDAs are created/destroyed when VM's are created/destroyed, the
default role should not be more privileged than this. The VDA
class is also likely, in the future, to have functions which
attempt to remediate any issues the VDA is experiencing, which
would qualify these future calls as potentially requiring at
least `VM Operator` privilege, which is required if the call
is "related to starting, stopping, pausing etc VMs or
otherwise accessing/manipulating VMs". Therefore, the `VM
Operator` role is chosen as the default for the VDA class.

Initially, one may think that the get-status and get-log-report
calls must only require a `Read only` user, but these calls will
actually contact the VDA inside the VM to get this information,
instead of just reading from the database, so the `Read only`
role is not privileged enough to execute them, therefore they
should also require at least `VM Operator` privilege. The calls
are already using the `VM Operator` role, so no changes were
needed in this commit for that part.

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>