VDI.get_nbd_info: make it more robust
To avoid returning an error when one of the DB references become invalid
while this function is running - these references will simply be skipped
and not included in the final list of URIs. This could happen due to
concurrent operations manipulating the PBDs of the SR, for example.

It is probably much more user friendly to skip these references that
become invalid, instead of returning an obscure error about an invalid
DB reference.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>