Xen_api_lwt_unix: close connections in rpc function
Close the connection that we open in the rpc function returned by
Xen_api_lwt_unix.make. Since the Xen_api_lwt_unix module itself does not
expose functions for creating, using, and closing connections, the user
would was not able to close the connections created by this rpc
function. Previously it opened a new connection for every request that
it did not close - this resulted in leaking open connections on the
client side and open file descriptors on the xapi side, which can cause
the client and/or xapi to fail after a large number of requests.

With this commit the behaviour of the Lwt bindings will be the same as
the behaviour of the Unix bindings in the xapi-client ocamlfind package
provided by xapi: a new connection will be opened and closed for each
XenAPI call.

I've added the call to disconnect to the rpc function returned by the
Lwt library, and not directly to the rpc function in lib/xen_api.ml,
because the core library in lib/ supports reusing connections for
multiple RPC calls, therefore that rpc function should not close the
connection.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>