CA-262911: Use poweroff rather than halt
xenopsd was changed to send poweroff instead of halt for newer guests
with commit 7a71ef8ac5e8 ("CA-217805: Prefer asking newer guests to
poweroff rather than halt").  Unfortunately there are some existing
guests (NetScaler/FreeBSD 8.4) that actually do halt when asked to. This
is not correct because the toolstack has been told to shutdown the VM,
not halt it. This is only seen with upstream QEMU because qemu-trad had
a bug in the emulated serial controller that caused FreeBSD to go
straight from halt to shutdown so this halting behavior was never
noticed.

Since all known guests support being asked to poweroff, send this
instead of halt.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>