CP-23741: Do not migrate VDAs on VM crosspool migrate
To avoid complex changes to the import/export functionality,
which is used in crosspool migration, we have opted to change
the design w.r.t. VDA migration; xapi will now only delete
the VDAs on the source but WILL NOT copy them over to the
destination pool (for the migrated VM and its snapshots).

The creation of the VDAs on the destination pool will be done by
the VDA detection mechanism (once implemented). See the REQ-330
design for more details.

This resolves the possible issue described in commit
559a64ec7a5e54d088902ef5bcabaf9635c33cf0.

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>