CP-24262: Do now wait on both handler threads when running the proxy
Fixes a file descriptor leak: By switching to `Lwt.choose` in place of
`Lwt.join` we make sure that we are not waiting forever for one of the
two threads to die after a connection has closed one but not the other.

In addition the code is refactored so that all the socked are always
wrapped in a context manager and closed in the same functions that
creates them.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>