CA-261065: use different datapaths for different connections
This commit fixes a bug where multiple clients connecting to /
disconnecting from the samve VDI in parallel interfere with each other.

We have seen clients hanging during connection, and also error messages
like
"Caught Storage_interface.Internal_error("Vdi_automaton.Bad_transition(0, 2)")"
on the server side.

The problem is that we use the pid here to construct the datapath id,
but we have Lwt threads here, so all the threads are running in the same
process and will have the same pid, and therefore all the clients
connecting to the same VDI will have the same datapath id. However,
different connections to the same VDI should be represented by different
datapaths and therefore different id strings - in xapi, the datapath
*is* actually this id string.

This commit fixes the issue by making sure that the different Lwt
threads handling the different clients use different datapaths. There
are many ways to do this, here I used the Uuidm library to generate a
random UUID for each connection.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>