CA-264331: Remove unused shutdown ack timeout setting
This setting was previously configurable from the `xapi.conf` file,
when xenopsd was still a part of xapi. When they split, the setting
was not transferred and nor was it wired up to be used by xenopsd.

This commit removes the setting from xapi.

A corresponding commit in xenopsd adds it there and makes it
configurable in `xenopsd.conf`.

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>