Only consider VDIs attached to non-snapshot VMs as attached in the VDI op checks
To fix CA-265050, where some CBT ops were disallowed on VDI snapshots
that were attached to VM snapshots.

This happened because the VM.checkpoint XenAPI function, which is called
when one takes a "disk and memory snapshot" from XenCenter, saves the
currently_attached fields of the snapshotted VM's VBDs in the
currently_attached fields of the VM snapshot's VBDs. This is different
from how this field is normally used: this field becomes true when the
VBDs actually get attached, and when the VM is shut down, the
currently_attached fields of the VBDs become false.

This confused the VDI checks in Xapi_vdi.check_operation_error, they
thought that since currently_attached is true, the VDI is actually
attached - which is usually true, but not in this case. In this case,
the VBDs just link the VDIs to a snapshot VM, but don't represent an
actual physical attachment of the VDI in SM. As a result, the checks
disallowed VDI.export_changed_blocks, and VDI.data_destroy on these
VDIs.

I had to update some of the existing tests because now only VDIs that
are attached to an actual non-snapshot VM are considered attached.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>