Xapi_vdi.check_operation error: don't fail due to invalid references
Use safe list functions that ignore exceptions due to invalid references
in the newly-added filtering. This can happen when some references to
other XenAPI classes become (or are already) invalid while the checks
are running, and causes an exception in the message forwarding layer
which calls this function to check for errors, which will prevent the
actual XenAPI function from being run. See CA-255128 for an example of
this kind of issue.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>