CP-24360: Initialise VM.platform['device-model'] values - Stage 1
The HVM templates of a host shall have a `VM.platform["device-model"]`
field containing the default device model according to the QEMU feature
release stages.

When a VM is created from a template, if `VM.platform["device-model"]`
is not set, then it shall be initialized from the default device model
profile.

Signed-off-by: Liang Dai <liang.dai1@citrix.com>