Xapi_vdi.data_destroy: don't call update_allowed_operations
Because it is already updated in the message forwarding layer by
unmark_vdi called from with_sr_andor_vdi, and we use that function for
data_destroy.

Also, add a unit test to verify that this is actually true. In this test
we have to go call the client instead of the implementation, to ensure
that we go through the message forwarding layer.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>