CA-265161: Enable default values for enums
This means that when we extend enums in the datamodel, older hosts
will be able to unmarshal them (although they will still not know
what to do with them). This patch causes marshalling of an `unknown
to be an error in all cases, although unmarshalling is OK.

I'd like this reviewed carefully, since the requirement is that we
fail 'safe' on encountering an unknown enum.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>