Allow intra-pool migration when CBT-enabled VDIs are not moved
By moving the check from the message forwarding layer to the
implementation, because in message forwarding we do not know the VDI to
SR mapping.

I added a check to Xapi_vm_migrate.migrate_send and assert_can_migrate,
checking that no CBT-enabled VDI is mapped to an SR different from its
current one. This is the only check that is done, whether the migration
is cross-pool or intra-pool. I assume the existing checks disallow
mapping a VDI to its current SR in case of cross pool migration, because
this pool's SR is probably not accessible to the remote pool. If this is
true, then cross-pool migration of VMs with CBT-enabled VDIs will never
be allowed, as expected, because all the VDIs have to be mapped to a
different SR.

In assert_can_migrate, I've added the checks for CBT-enabled VDIs to the
end, because unlike the errors checked first, these errors are
recoverable - the user can just disable CBT on the affected VDIs.

The other errors may not be possible to fix, therefore those checks
should take precedence over the CBT checks.

Signed-off-by: Iglói Gábor <gabor.igloi@citrix.com>