Disallow data_destroy on VDIs that are linked to a VM (VDIs with VBDs)
Previously data_destroy was allowed on VDIs linked to a VM snapshot. In
order to maintain the invariant that VBDs only point to valid (not
cbt_metadata) VDIs, in data_destroy we remove all the VBDs linked to the
VDI, just like for destroy. But as a result, running data_destroy a VDI
of a VM snapshot produced an invalid, unbootable VM snapshot missing its
boot disk that was still visible to the user. The same thing would
happen in case of a stopped VM, whose VBDs are not attached.

Now users have to remove the VM snapshot (or the VBD) first, before
data_destroying the VDIs. The same applies to stopped VMs: the VBD has
to be removed first for data_destroy to succeed.

When implementing the new check, I assumed that all the VBDs of the VDI
(the VBDs in my_vbd_records) are linked to an actual, non-null VM ref.
In case the VM ref is not valid or null, that should be only temporary,
because the next DB GC pass would clean up the invalid VBD. So I ignored
this edge case in the check, and did not filter out VDIs linked to null
/ invalid VMs, because the DB GC should fix this anyway, and then the
check will pass. VBDs cannot be created with null VM references, so this
situation normally shouldn't happen.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>