CA-266936: pciutils: use the file dump in a temp file if possible
Rarely xapi is booting and updating the PCI records while the data in
sysfs is being written/changed. This invalidated the `access` structure
and caused a segmentation fault.

By using the `dump` access, that in practice means dumping the
equivalent of `pci -x` in a temporary file (see man pciutils) we can
make sure that the data is not changing under our feet.

This seems to have resolved the segmentation fault, at the price of
potentially having partial information about the pci devices on the
host. This did not seem to affect system tests, and in the worst case
scenario, after a toolstack restart all the pci devices are detected.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>