Fix the SDK
The SDK is quite limited in the types it supports - in particular it
sees to have a problem with a (_ Ref * Enum) Map. This has been worked
around in the current_operations code by taking advantage of the fact
that a Ref is really just a string.

Additionally, allowed_and_current_operations is a useful function
used in the other object definitions to generate the 2 fields, and
contains this workaround. So this commit replaces the explicit
definitions of those fields with a call to this helper function.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>