CP-24774: use QMP event XEN_PLATFORM_PV_DRIVER_INFO to set HVM Linux feature flags
Before the QMP Event handler in the qemu-upstream backend, we relied on the
xenops_server_xen.ml function `maybe_update_pv_drivers_detected` to detect the
PV drivers connected state 4 to decide when to set the HVM Linux feature flags
in xenstore.

This, however, wasn't completely backwards-compatible with qemu-trad because
qemu-trad would set these flags when the Xen hypervisor platform driver inside
the HVM Linux guest initialised, which is earlier and re-occurs during migration.
This difference in behaviour caused detectable delay-related changes during guest
startup and lack of these feature flags after live migration.

After the QMP Event handler was added to the qemu-upstream backend in xenopsd,
it's possible to fix this issue by relying on the XEN_PLATFORM_PV_DRIVER_INFO
QMP event available in qemu-upstream, which is emitted at exactly the same
point as in qemu-trad, when the HVM Linux guest's PV-aware kernel writes to the
Xen hypervisor platform PCI device.

The improvement above is also an opportunity to remove the public function
`maybe_write_pv_feature_flags`, which was only added to permit it to be called
from the function `maybe_update_pv_drivers_detected` in xenops_server_xen.ml.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>