Test_vdi_cbt: update existing data_destroy tests after behaviour changes
Now data_destroy will first wait for all the VBDs to be removed, so it's
not important to test that this happens. However, removing the VBDs is
still useful to correct the race condition where the VBD gets created
while data_destroy is running, but that's very hard to unit test.

Signed-off-by: Iglói Gábor <gabor.igloi@citrix.com>