CA-265683: Use a sub-task for VM.assert_can_migrate inside VM.migrate_send
VM.assert_can_migrate may now forward a VM.assert_can_migrate_sender call to
the sending host, but it used the same task that was created for the
VM.migrate_send call. This caused xapi to mark the entire migrate_send task as
completed as soon as the assert_can_migrate_sender call finished. Using a
sub-task avoids this.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>