CA-267831: Remove bad assert_can_boot_here call from intra-pool SXM
Instead of `reserve_memory_for_vm` we call `allocate_vm_to_host`, because the
former calls `assert_can_boot_here` without taking into account that the
storage is going to be migrated. It is an unnecessary check anyway, because it
is preceded by assert_can_migrate, which does similar checks (but the right
ones).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>