Storage_migrate.receive_start,copy: filter out cbt_metadata VDIs
Filter out VDIs of type cbt_metadata that don't have any actual data,
when finding the nearest VDI in these functions. This is to avoid
accidentally using a cbt_metadata VDI without data. The
VDI.similar_content call already does this filtering when we look for
similar local VDIs, and this commit extends the filtering to the remote
VDIs that we compare the similar VDIs to.

The content_id of all the cbt_metadata VDIs is set to the same constant,
so this change prevents accidentally matching two of these to each
other when we try to find the nearest VDI.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>