CA-267954: prevent too strict association of QMP.event_data type in xenopsd
Matching the QMP.event_data options in xenopsd without a wildcard binds
xenopsd too strictly to a specific version of the ocaml-qmp library. This
causes xenopsd to fail to compile unnecessarily when updating the ocaml-qmp
library to support new QMP.event_data options. This patch adds this missing
wildcard, so that ocaml-qmp can be improved without failing xenopsd.

We do not print a debug message when unhandled events are received because
all incoming events are already printed, and the unhandled ones are easy to
infer by subtracting these from the printed ones.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>