Xapi_vdi.snapshot: remove logic setting cbt_enabled in the DB
We should not try to guess here whether CBT is enabled or not on the
snapshotted VDIs. SM should be responsible for updating this VDI field
in the DB, instead of xapi trying to guess and keep track of it, as SM
is in the best position to know when CBT is enabled, and it is already
setting this field during SR scan.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>