Test_vdi_cbt: don't test how VDI.clone & snapshot updates VDI.cbt_enabled
Updating this field is the responsibility of the SM layer. Therefore we
cannot currently unit test this any more - we should add corresponding
quicktests instead to replace this unit test.

Another option is to mock the SMAPIv1 layer to unit test this - however,
it may not be worth doing this, as SMAPIv1 may be superseded in the
future.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>