CP-24805: set the default VM.platform device-model profile during upgrade
If the `VM.platform["device-model"] field of a HVM VM is not set, then
it shall be initialised from the fallback device model profile when
a host is upgraded to a newer version of XenServer, using the Xapi
database upgrade mechanism. During the upgrade, the database initialises
this value using the fallback device model profile value for the
XenServer version that the host is being upgraded from.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>