CP-24687: Simplify the pool valid operations code
This code, while being more flexible, has unnecessary extra calls
and can be defined more succintly for the current requirements of
checking which pool operations are allowed.

For the moment, we make the code cleaner by taking advantage of the
fact that no concurrent pool operations are allowed (with the
existing set of pool operations, which are HA enable, HA disable and
cluster creation).

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>