cp-24831: persist qemu backend profile in xenopsd VMExtra.persistent_t
The VMExtra.persistent_t type is a xenopsd facility that persists VM metadata
values throughout the lifetime of a VM across all VM lifecycle operations
involving xenopsd: start, suspend, resume, migrate, until shutdown. This is an
ideal place to robustly persist the profile that is used to select the qemu
backend throughout the lifetime of a HVM VM, because of the profile of a VM
does not change after it has started. This is a much better approach than
relying on a mutable xenstore key to store the profile, as is currently done.

This patch introduces the profile member to this type, and sets its value to
be the same profile used to select the qemu backend for HVM VMs. Subsequent
patches will select the backend using this persistent profile instead of
using the mutable xenstore key.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>