Fix mutex in VUSB.create
The mutex is to protect the pre-checks, which limit the number of VUSBs per VM,
and per USB group, in the case that multiple VSUB.create calls come in. The
critical section must start before the `Db.VM.get_VUSBs` call for this to be
thread safe.

Also note that the final couple of lines of the function are part of the second
case in the `ha_restart_priority` match, even though the indentation suggests
otherwise. This is actually (luckily) alright, because the first case raises an
exception.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>