CP-25021: Generate automatically sets of records in the C and C# SDK.
Calls returning sets of records are a rare occurence in the API and so far their
deserialisation functions and types have been added to the C/C# SDK manually and
on demand. As of this commit they are generated automatically so they are
available for any calls that need them.

Signed-off-by: Konstantina Chremmou <konstantina.chremmou@citrix.com>