CA-270642, CA-265691: Copy metrics from snapshot when reverting
Do not copy the metrics reference from the `snapshot_metadata`, because this
comes from the original VM (the one that the snapshot was taken of), and may
have been destroyed meanwhile. The snapshot itself has a copy of the VM_metrics
record of the original VM at the time the snapshot was taken, so copy that
when restoring the VM.

This applies to snapshots as well as checkpoints.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>