CA-263119: Use adjusted maxmem in execute_action
In CA-263119 we had all HVM VMs declared as uncooperative, and squeezed failing to meet its target. The domains were not really uncooperative (if you manually wrote a lower target in xenstore they adjusted their memory usage), but the target was not actually achievable because maximum memory was set lower than the target by about 1 MiB. From the VM's point of view it seemed it has done as much ballooning as possible.
Squeezed accounted for the HVM offset when writing maxmem, but not when reading maxmem, which caused the safety check (of only writing targets <= maxmem) not work properly.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>