[CA-271014] Only get local SRs on the physical utilisation thread
This greatly improves the performance of the pool master in a large pool. The current loop means that every 2 minutes each host in the pool queries, for each PBD (regardless of which host it's on), whether it's currently attached, and its SR. It then updates the xapi database with the RRDD physical utilisation value for those SRs it has the data. As an optimisation therefore, this commit adds a new method to only get PBDs on the calling host, and then the SRs for those will be those that it can update the RRDD data for.

The issue with this physical utilisation thread is the overhead on the pool master, which must handle all the database queries. This is now only one query to get the PBDs currently attached on the calling host, plus one more for each PBD currently plugged on localhost (to get its SR). It was previously one query to get all PBDs on the pool (regardless of host), plus an extra two queries for each PBD to get its currently_attached and then its SR.

If a pool has n hosts with m PBDs each (all currently attached), then we previously had n hosts making 2nm+1 database queries each every 2 minutes, and we now have n hosts making 1+m database queries each every 2 minutes.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>