quicktest: allow filtering by SR name
By default some tests are run on all SRs, or only on the first SR.

When debugging it is useful to run tests only on a specific SR (e.g. GFS2),
it is also useful to filter this by a specific SR name rather than type because
you could have multiple SRs attached of the same type, and some of them might be
broken.

When doing component level "medium" testing it'd also be useful to exercise all
operations on the specific SR we are testing the integration with.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>