cp-24831: persist qemu backend profile when importing VM metadata in migration
During the execution of a VM_migrate operation, xenopsd transmits the VM
metadata using the function `Remote.VM.import_metadata`, which is executed on
the destination host. This function will eventually call `set_internal_state`
on the destination host, which recomputes some fields the VMExtra metadata for
the VM about to be received as a result of the migration, before persisting it
in the destination host.

This patch recomputes the profile field in VMExtra as well before persisting
this metadata, which is necessary to guarantee that the received metadata will
contain a profile consistent with the destination host. This is specially
significant if the destination host is a newer version of xenopsd, which could
associate different qemu backend profiles for the received VM metadata.

An extra line of debugging for each vm migrate operation is added, to permit to
check which VMExtra fields are being persisted on the new host.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>