CP-25468: qemu-wrapper: Enable seccomp sandboxing
Deny access to a variety of system calls for QEMU by turning on seccomp
sandboxing (for example, fork and execve). The syscalls for changing
privileges (e.g. setuid) are explicitly allowed, because otherwise it is
not possible to later become an unprivileged user. This should not be a
problem, because the unprivileged user would not be able to call
setuid(), etc. anyway.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>