CP-25570: connect to VNC console in qemu-upstream VMs using a Unix socket
instead of using a network connection. This makes it easier to separate
qemu-upstream in its own network namespace. A nice side-effect is that xenopsd
knows the Unix path of the socket that was used when starting the qemu-upstream
process and therefore QMP queries to qemu-upstream are not necessary anymore,
simplifying the qemu-upstream backend.

qemu-trad and PV VMs are not affected and continue to use a network connection
for the VNC console.

This change depends on replacing the parameter for the VNC console from a
network port to a Unix socket path when starting qemu-upstream -- this will
occur in the next commit.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>