CA-272126: Don't get NVidia vGPU types from XML file for now.
This bypasses the changes from b66e46cc, until low-level bits have been fixed.

This is the "smallest" fix for this issue. We already have a fallback to the
old method for the case that the XML is invalid. We'll now take that path even
when the XML did give us some valid vGPU types.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>