CP-25570: use a unix socket for the vnc console when starting qemu-upstream
using the same matching vnc_socket_path of the previous commit. After this
change, qemu-upstream will be started using

>  -vnc unix:/var/run/xen/vnc-<domid>,lock-key-sync=off
>    A replacement for qemu-trad's -vncunused -vnc 127.0.0.1:1.

instead of

>  -vnc 127.0.0.1:1,to=1000,lock-key-sync=off
>    Equivalent to qemu-trad's -vncunused -vnc 127.0.0.1:1.

This change depends on removing the corresponding vnc console parameter
transformation in the qemu-upstream wrapper. The comment for the
lock-key-sync=off option was moved from this wrapper.

The qemu-trad codepath is not affected.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>