xapi.conf server_cert_path not respected, rework+
In this change, the server_cert_path variable in xapi.conf is used by xapi as arguement to pass to xapissl script to have the ssl configuration file generated to use this configured value.
It includes changes in xapi to passthrough the arguments and xapissl script to accept and use the value from argument.

Signed-off-by: Yarsin He <yarsin.he@citrix.com>