CA-274113: Create hosts as enabled on the slave
The master's cluster_host object is created as enabled,
however the slaves' always got created as disabled.

Now that we are actually checking the prerequisite
(that cluster_host needs to be enabled in order to plug a PBD)
this causes random failures in the BVT depending on which host it happens to
start the VM on.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>