CP-25873: qemu-wrapper: Create tap devices before starting QEMU
Commit 34d357ac25b5 ("qemu-wrapper: Call unshare() before executing
QEMU") broke emulated NIC devices because the tap devices get opened in
a separate network namespace. Create the tap devices before starting
QEMU and pass them as open file descriptors. Implement it as a quick
workaround in qemu-wrapper for now to unblock the build, but this should
eventually be implemented somewhere else.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>