CA-274107: Forward cluster host enable/disable to correct host
The cluster host operations must all run on the host in which the
operation is being called for because they must communicate with
the host's local xapi-clusterd daemon. An assertion is also added
in the local implementations of the cluster host operations so that
this condition is always checked before executing a cluster host
opeation.

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>