CP-20544: Initialize coverage dispatcher
See `CP-22945: runtime coverage dumping support for Xcp_service` in `xapi-idl`.
This defines a coverage dispatcher that is a no-op by default, enabled only when
`xapi-idl` is compiled in a special way.

XAPI is the only daemon in the system that should be initializing it,
when runtime coverage dumping is enabled this will create a new thread listening
on a message queue, but for all normal builds this will be a no-op.

(all the conditional compilation is done in `xapi-idl` instead of all daemons
 needing coverage).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>