CA-273306: do not assert the VM power_state during VM.checkpoint
The `xapi_vgpu.create` function expects that the passed vm record is always
in the Halted power_state. However, during a VM.checkpoint, the power_state
of the VM is Suspended and causes the internal Halted assertion to fail.

This occurs because during a VM.{checkpoint,snapshot,clone} operation, a
VGPU record is implicitly created in the xapi db when it is copied from the
original VM. In VM.checkpoint specifically, this copy occurs while the VM is
in the Suspended power_state. Measured power_states during tests:

| Operation                 | measured VM power_state during VGPU.copy |
+---------------------------+------------------------------------------+
| VM.checkpoint             | Suspended                                |
| VM.snapshot               | Halted                                   |
| VM.revert_from_checkpoint | Halted                                   |
| VM.clone/copy             | Halted                                   |

There's no special reason to restrict the VGPU copy operation only to the
Halted power_state. Other power_states, such as Suspended, are acceptable.

The original `xapi_vgpu.create` function signature cannot be changed because
it is autogenerated from the xapi datamodel, so this commit creates an
internal `create'` function that accepts a boolean determining if the original
power_state check for Halted should be executed or not. This parameter is
then used by the existing datamodel-derived functions:
* `xapi_vgpu.create` function behaves exactly as before, and it continues to
assert the Halted power_state at all times.
* The `xapi_vgpu.copy` function, called during VM snapshot, checkpoint and
clone, does not execute the assertion anymore and therefore fixes the issue.

A priori, fixing the issue could be done by just removing the assertion in the
`xapi_vgpu.create` function. The only side effect (apart from fixing the issue)
is that it would be possible to use the XenAPI directly to call VGPU.create at
any time -- `xe vgpu-create` would not enforce the VM to be in the Halted
state anymore. This would only affect the Xapi db, creating a slightly
inconsistent view of the properties of the VM until the VGPU is actually
attached to the VM in the next VM.start.

However, in this case XenCenter will show this inconsistent view of the running
VM, which will not have a VGPU attached but may show one in its 'properties',
which could cause some confusion to the user, so it's probably better to keep
the create function (as accessed from the XenAPI) protected by the assertion.

Therefore, after this change, `xe vgpu-create` will still return an error if
the VM is not Halted:

  # xe vgpu-create ...
  You attempted an operation on a VM that was not in an appropriate power state
  expected: halted
  actual: running

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>