CA-270640: Avoid leaking reader threads
In case of errors, the monadic bind will "shortcircuit" potentially
leaking reader threads blocked and waiting to be woken up.

In the result collection thread this can be fixed by making sure that
we are always iterating on all of them.

This makes the code slightly less clean and relies on the fact that
let bound results are eagerly computed before the bind operator is
applied to process the final value.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>