Xapi_sr.update_vdis: correctly set VDI.sharable field
This update_vdis function is called from places like SR.scan and
VDI.update.

Now that we have SMAPIv3 plugins that are not backed by the database,
xapi really needs to check the returned vdi_info records and update the
database if necessary. Previously the SMAPIv1 plugins updated xapi's
database, so the whole code in update_vdis did nothing. But now with
SMAPIv3 plugins, we have to ensure that we propagate the information
from the vdi_info records returned by the SM layer into the database,
because it may not be in the database, and it may be different from
what's in the database.

This change does not break SMAPIv1 plugins, because the
vdi_info_of_vdi_rec function in storage_access.ml fetches the correct
VDI.sharable field from xapi's database, so in case of SMAPIv1 plugins,
this should be a no-op. This vdi_info_of_vdi_rec is the function we use
in SR.scan and VDI.stat in our SMAPIv1<->SMAPIv2 translator in
storage_access.ml to return a list of vdi_info's, or a single vdi_info,
respectively, from xapi's database.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>